refactor(capabilities): split trampoline.rs into a trampoline/ submodule
`trampoline.rs` (698 lines) becomes `trampoline/{mod.rs, state.rs,
replace.rs}` split along its real seams. `state.rs` holds the
`WasmTrampolineConfig` and `WasmTrampoline` struct definitions;
`replace.rs` holds the inherent `impl WasmTrampoline` (spawn_sibling,
resolve_replace_target, handle_replace); `mod.rs` retains the bridge
block with the `#[actor] impl NativeActor` lifecycle and the tests.

Both `state` and `replace` are declared at the `trampoline` module
level (file-root siblings of the bridge block, following the
`audio/decode` precedent), not inside the `#[bridge]` token stream —
file-backed `mod` declarations inside proc-macro input are an unstable
Rust feature (E0658). Fields on `WasmTrampoline` are `pub(super)`
(visible within the `trampoline` module subtree) so the sibling
`replace` module and the inline `mod native` block can access them
without widening past the crate.

No public-API change: `WasmTrampoline` and `WasmTrampolineConfig`
continue to resolve through the existing `lib.rs` re-exports. No
kinds move; no behavior change.

Closes #2169
